==> movie_recsys/05_data_preprocessing_CTM.ipynb <==
## 05_data_preprocessing_CTM.ipynb

Description: 
- build dataset for collaborative topic model (CTM), combining movieLens dataset and scripts dataset

External Dependencies:
- MovieLens links.csv from "..\\database\\dataset_movieLens\\links.csv"
- MovieLens ratings.csv from "..\\database\\dataset_movieLens\\ratings.csv"
- springfield_movie_scripts.csv from "database\dataset_film_scripts\springfield_movie_scripts.csv"

Returns:
- pd.DataFrame saved to "data_preprocessing_out\\df_processed_CTM.csv"

import pandas as pd
import numpy as np

# load in cleaned movie scripts dataset
df_script = pd.read_csv(
    "..\\database\\dataset_film_scripts\\springfield_movie_scripts.csv", index_col=[0]
)
df_script = df_script.drop(
    ["script_text", "springfield_link", "tmdb_poster_link", "imdb_link"], axis=1
)
df_script["recsys_id"] = df_script.index
print(df_script.head())

# load in movieLens links and ratings datasets
df_movielens = pd.read_csv("..\\database\\dataset_movieLens\\links.csv")
print(df_movielens.head())

df_movielens_ratings = pd.read_csv("..\\database\\dataset_movieLens\\ratings.csv")
print(df_movielens_ratings.head())

# join movieLens dataset with scripts dataset
df_joined = df_script.join(
    df_movielens.dropna().set_index("tmdbId"), how="left", on="tmdb_id"
)

# drop duplicates and missing movieIds
df_joined = df_joined.drop_duplicates(subset="tmdb_id")
df_joined = df_joined.dropna(subset="movieId")

df_joined.head()

# filter out movies from ratings matrix that are not in script database
# takes about 3.5 minutes to run
unique_movielens_ids = df_joined["movieId"].unique()[1:]
unique_movielens_ids = np.sort(unique_movielens_ids.astype(int))
movieId = np.array(df_movielens_ratings["movieId"])

bool_mask = [True if j in unique_movielens_ids else False for j in movieId]
df_movielens_ratings = df_movielens_ratings.loc[bool_mask]

# drop all users from ratings matrix that rated less than n_films
n_films = 2
unique_movielens_users = np.array(df_movielens_ratings["userId"].value_counts().index)
num_ratings_per_user = np.array(df_movielens_ratings["userId"].value_counts())
userId = np.array(df_movielens_ratings["userId"])

users_drop = unique_movielens_users[num_ratings_per_user < n_films]
bool_mask = [False if j in users_drop else True for j in userId]
df_movielens_ratings = df_movielens_ratings.loc[bool_mask]


df_movielens_ratings.head()

df_final = df_movielens_ratings.join(
    df_joined.set_index("movieId"), how="left", on="movieId"
)
df_final = df_final.dropna()
df_final["recsys_id"] = df_final["recsys_id"].astype("int")

df_final.head()

df_final[["userId", "recsys_id", "rating"]].to_csv(
    "data_preprocessing_out\\df_processed_CTM.csv"
)
